refactor: infra 코드의 패키지 변경, 애너테이션 컨벤션 적용
