+++ src/test/java/com/woowacourse/pelotonbackend/infra/login/KakaoAPIServiceTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pelotonbackend.member.infra;
+package com.woowacourse.pelotonbackend.infra.login;
 
 import static com.woowacourse.pelotonbackend.member.domain.LoginFixture.*;
 import static com.woowacourse.pelotonbackend.member.domain.MemberFixture.*;
@@ -21,7 +21,6 @@ import org.springframework.web.util.DefaultUriBuilderFactory;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.woowacourse.pelotonbackend.infra.login.KakaoAPIService;
 import com.woowacourse.pelotonbackend.support.JwtTokenProvider;
 import com.woowacourse.pelotonbackend.support.dto.JwtTokenResponse;
 import okhttp3.mockwebserver.Dispatcher;
